requirements.txt
python-dotenv
sounddevice
numpy
simpleaudio
openai
speechrecognition
pocketsphinx
pyttsx3
pyaudio